+++ src/pages/doctor/Appointments.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import "./appointments.scss"
+
 
 const Appointments = () => {
   return (
