+++ src/pages/doctor/DoctorProfile.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import "./doctorProfile.scss"
+
 
 const DoctorProfile = () => {
   return (
