+++ src/pages/doctor/DoctorSettings.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import "./doctorSettings.scss"
+
 
 const DoctorSettings = () => {
   return (
