+++ src/pages/polyclinic/Account.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import "./account.scss"
+
 
 const Account = () => {
   return (
