--- src/pages/polyclinic/doctors/newDoctorRegister/NewDoctorRegister.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import "./newDoctorRegister.scss"
-
-const NewDoctorRegister = () => {
-  return (
-    <div>NewDoctorRegister</div>
-  )
-}
-
-export default NewDoctorRegister
